Moved TypeLegend and TypeStyleRepresentation to layer-schema-types.ts
