+++ packages/geoview-core/src/geo/layer/gv-layers/abstract-gv-layer.ts
@@ -32,12 +32,16 @@ import type {
   codedValueType,
   rangeDomainType,
 } from '@/api/types/map-schema-types';
-import type { TypeLayerMetadataFields, TypeGeoviewLayerType, TypeMetadataEsriDynamicLayer } from '@/api/types/layer-schema-types';
+import type {
+  TypeLayerMetadataFields,
+  TypeGeoviewLayerType,
+  TypeMetadataEsriDynamicLayer,
+  TypeLegend,
+} from '@/api/types/layer-schema-types';
 import type { GeoViewError } from '@/core/exceptions/geoview-exceptions';
 import { LayerFailedToLoadError, LayerImageFailedToLoadError } from '@/core/exceptions/geoview-exceptions';
 import type { TypeLegendItem } from '@/core/components/layers/types';
 import { GeoviewRenderer, type TypeStyleProcessorOptions } from '@/geo/utils/renderer/geoview-renderer';
-import type { TypeLegend } from '@/core/stores/states/layer-state';
 import { AbstractBaseGVLayer, type LayerBaseDelegate, type LayerBaseEvent } from '@/geo/layer/gv-layers/abstract-base-layer';
 import type { SnackbarType } from '@/core/utils/notifications';
 import { formatError, NotImplementedError, NotSupportedError } from '@/core/exceptions/core-exceptions';
