+++ packages/geoview-core/src/geo/layer/gv-layers/raster/gv-image-static.ts
@@ -8,11 +8,10 @@ import { logger } from '@/core/utils/logger';
 import type { ImageStaticLayerEntryConfig } from '@/api/config/validation-classes/raster-validation-classes/image-static-layer-entry-config';
 import { GeoviewRenderer } from '@/geo/utils/renderer/geoview-renderer';
 import { AbstractGVRaster } from '@/geo/layer/gv-layers/raster/abstract-gv-raster';
-import type { TypeLegend } from '@/core/stores/states/layer-state';
 import { Projection } from '@/geo/utils/projection';
 import { GeoUtilities } from '@/geo/utils/utilities';
 import { Fetch } from '@/core/utils/fetch-helper';
-import { CONST_LAYER_TYPES } from '@/api/types/layer-schema-types';
+import { CONST_LAYER_TYPES, type TypeLegend } from '@/api/types/layer-schema-types';
 
 /**
  * Manages an Image static layer.
