+++ packages/geoview-core/src/geo/layer/gv-layers/tile/gv-geotiff.ts
@@ -3,9 +3,8 @@ import type GeoTIFFSource from 'ol/source/GeoTIFF';
 import type { Extent } from 'ol/extent';
 import type { Projection as OLProjection } from 'ol/proj';
 
-import { CONST_LAYER_TYPES } from '@/api/types/layer-schema-types';
+import { CONST_LAYER_TYPES, type TypeLegend } from '@/api/types/layer-schema-types';
 import type { GeoTIFFLayerEntryConfig } from '@/api/config/validation-classes/raster-validation-classes/geotiff-layer-entry-config';
-import type { TypeLegend } from '@/core/stores/states/layer-state';
 import { AbstractGVTile } from '@/geo/layer/gv-layers/tile/abstract-gv-tile';
 import { GeoUtilities } from '@/geo/utils/utilities';
 import { Projection } from '@/geo/utils/projection';
