+++ packages/geoview-core/src/geo/layer/gv-layers/tile/gv-wmts.ts
@@ -8,12 +8,11 @@ import type {
   OgcWmtsLayerEntryConfig,
   TypeMetadataWMTSLayer,
 } from '@/api/config/validation-classes/raster-validation-classes/ogc-wmts-layer-entry-config';
-import type { TypeLegend } from '@/core/stores/states/layer-state';
 import { AbstractGVTile } from '@/geo/layer/gv-layers/tile/abstract-gv-tile';
 import { GeoUtilities } from '@/geo/utils/utilities';
 import { Projection } from '@/geo/utils/projection';
 import { Fetch } from '@/core/utils/fetch-helper';
-import { CONST_LAYER_TYPES } from '@/api/types/layer-schema-types';
+import { CONST_LAYER_TYPES, type TypeLegend } from '@/api/types/layer-schema-types';
 import { logger } from '@/core/utils/logger';
 
 /**
