+++ packages/geoview-core/src/geo/utils/utilities.ts
@@ -24,17 +24,17 @@ import { CONST_LAYER_TYPES, validVectorLayerLegendTypes } from '@/api/types/laye
 import type { TypeMapMouseInfo, TypeOutfields, TypeStyleGeometry, TypeValidMapProjectionCodes } from '@/api/types/map-schema-types';
 import type {
   TypeGeoviewLayerType,
+  TypeLegend,
   TypeMetadataWMS,
   TypeMetadataWMSCapabilityLayer,
   TypeMetadataWMSRoot,
   TypeStylesWMS,
+  TypeVectorLayerStyles,
 } from '@/api/types/layer-schema-types';
 import type { TypeBasemapLayer } from '@/geo/layer/basemap/basemap-types';
 import { NetworkError, NotSupportedError, ResponseEmptyError } from '@/core/exceptions/core-exceptions';
 import type { TypeMetadataWMTS } from '@/api/config/validation-classes/raster-validation-classes/ogc-wmts-layer-entry-config';
-import type { TypeLegend } from '@/core/stores/states/layer-state';
 import type { TypeLegendItem, TypeLegendLayerItem } from '@/core/components/layers/types';
-import type { TypeVectorLayerStyles } from './renderer/geoview-renderer';
 
 // available layer types
 export const layerTypes = CONST_LAYER_TYPES;
